+++ CHANGELOG.md
@@ -47,6 +47,7 @@ TODO: check how to setup a "blog" section in the website for major releases, to 
 - **typing**: `Dict::from_object` now returns `Dict[str, object]` instead of `Dict[str, Any]` to avoid silencing errors with strict type checkers (e.g Pyrefly, Basedpyright) who forbid the use of `Any` in generic types. You can always use `typing::cast` if you need specific attribute access without runtime check AND no type checker errors.
 - **typing**: `Result` and `Option` method `transpose` give less false positives now.
 - **typing**: Thanks to covariance, `PyoIterator::flatten` should not give false positives anymore, since it seems there's no need anymore to manually maintain various overloads for nested `PyoIterator` types.
+- **typing**: `PyoIterator::chain` now has overloads to return precise tuple types for up to 5 iterables, and handle covariance in all cases.
 
 ### 🐞 Bug fixes